DefaultUIModel: Reworked the receiving of radio messages.
The original implementation did not work because the API to receive
web socket messages was not correctly understood. The web socket
request yields a future that completes only if the underlying
connection terminates. This is now handled correctly, and an exchange
of radio messages between server and client works as desired at this
stage.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>