DefaultUIModel: Changed internal data model for radio playback state.
Rather than having one Var with the whole state, the data is now split
over multiple Vars for the single components, which are combined to
the radio state signal. This simplifies updates for single components,
since it is not necessary to deal with the Option and Try types.

The current radio source and the replacement source are now stored
using their ID values and resolved against the list of available radio
sources. This is necessary to handle some radio messages that refer to
sources only via their IDs.

Signed-off-by: Oliver Heger <oliver.heger@oliver-heger.de>